Refactor Nutzap profile helpers to break circular import
